d change to zu

+++ c/infosec/6. Integer Overflow/main.c
@@ -1,11 +1,11 @@
 #include <stdio.h>
 
 int main() {
-  unsigned int limit = 10;
+  size_t limit = 10;
 
   // what if replace greater operator (>) with greater and equal operator (>=)
   while (limit > 0) {
-    printf("limit: %d\n", limit);
+    printf("limit: %zu\n", limit);
     --limit;
   }
